Reorganized bevy-related modules
