+++ kayak_font/src/bevy/renderer/mod.rs
@@ -1,2 +1,5 @@
+mod extract;
 mod font_texture_cache;
+
+pub use extract::*;
 pub use font_texture_cache::*;
